added constant file and added more tests
