+++ src/applications/rated-disabilities/util/datadog-rum/initializeRealUserMonitoring.js
@@ -2,17 +2,7 @@ import { datadogRum } from '@datadog/browser-rum';
 
 import environment from '@department-of-veterans-affairs/platform-utilities/environment';
 
-const DISABLED_ENVIRONMENTS = ['localhost'];
-
-export const canUseRUM = () => {
-  const env = environment.vspEnvironment();
-
-  const alreadyInitialized = Boolean(window.DD_RUM?.getInitConfiguration());
-  const inTestEnv = window.Cypress || window.Mocha;
-  const inDisabledEnv = DISABLED_ENVIRONMENTS.includes(env);
-
-  return !alreadyInitialized && !inTestEnv && !inDisabledEnv;
-};
+import { useRUM } from '../../constants';
 
 // https://docs.datadoghq.com/real_user_monitoring/browser/#configuration
 const defaultRumSettings = {
@@ -40,12 +30,6 @@ const defaultRumSettings = {
 // Don't call this function if not logged in
 export default function initializeRealUserMonitoring(customRumSettings) {
   // Prevent RUM from re-initializing the SDK OR running on local/CI environments.
-
-  // This should only be set to `true` to enable RUM locally
-  // Otherwise this should be `false`
-  const overrideChecks = false;
-  const useRUM = canUseRUM() || overrideChecks;
-
   if (useRUM) {
     datadogRum.init({
       ...defaultRumSettings,
